sse: Improve SSE parser to be compliant
